+++ src/main/java/com/mageddo/dnsproxyserver/server/dns/DnsServerStarter.java
@@ -1,7 +1,7 @@
-package com.mageddo.dnsserver;
+package com.mageddo.dnsproxyserver.server.dns;
 
 import com.mageddo.dnsproxyserver.config.application.Configs;
-import com.mageddo.dnsproxyserver.server.dns.solver.docker.application.DpsContainerService;
+import com.mageddo.dnsserver.SimpleServer;
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
 
@@ -11,18 +11,16 @@ import javax.inject.Singleton;
 @Slf4j
 @Getter
 @Singleton
-public class ServerStarter {
+public class DnsServerStarter {
 
   private final SimpleServer server;
-  private final DpsContainerService dpsContainerService;
 
   @Inject
-  public ServerStarter(SimpleServer server, DpsContainerService dpsContainerService) {
+  public DnsServerStarter(SimpleServer server) {
     this.server = server;
-    this.dpsContainerService = dpsContainerService;
   }
 
-  public ServerStarter start() {
+  public DnsServerStarter start() {
     final var config = Configs.getInstance();
     final var port = config.getDnsServerPort();
     this.server.start(
