+++ src/main/java/com/mageddo/dnsserver/SimpleServer.java
@@ -1,5 +1,6 @@
 package com.mageddo.dnsserver;
 
+import com.mageddo.dnsproxyserver.server.dns.RequestHandlerDefault;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
